Add files via upload
An ideal steam turbine uses 6000 kg/h   of steam. The steam is superheated to 720°C   and 4000 kPa when supplied to the turbine. The exhaust temperature is 80°C, and the fluid leaving the condenser is assumed to be a saturated liquid. Determine the power generated by the cycle, the rate of heat rejected the amount of coal consumed if the boiler unit is assumed to be 100% effective while the LHV of coal is 30,000 kJ/kg, and the steam rate.